Solution to Check Triangle Validity

+++ Check Triangle Validity@@6788dd63490637cfc44e1b93/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float x,y,z;
-    scanf("%f %f %f",&x,&y,&z);
+    int x,y,z;
+    scanf("%d %d %d",&x,&y,&z);
     if(a+b>c && a+c>b && b+c>a){
         printf("Valid");
     }
